refactor(css/lints): Extract common functions
