+++ crates/swc_css_lints/src/lib.rs
@@ -1,6 +1,7 @@
 #![deny(clippy::all)]
 
 mod config;
+mod dataset;
 mod rule;
 mod rules;
